Reorganize microphysics folder into models module
