+++ external/fv3fit/fv3fit/emulation/layers/__init__.py
@@ -11,9 +11,9 @@ from .fields import (
     IncrementedFieldOutput,
     FieldInput,
     FieldOutput,
-    CombineInputs,
 )
 from .architecture import (
     MLPBlock,
     RNNBlock,
+    CombineInputs,
 )
