+++ external/fv3fit/fv3fit/emulation/layers/architecture.py
@@ -1,3 +1,4 @@
+from typing import Optional
 import tensorflow as tf
