+++ external/fv3fit/fv3fit/emulation/models/__init__.py
@@ -1,1 +1,1 @@
-from .models import Config
+from .microphysics import MicrophysicsConfig
